refactor:메모 읽기 전용으로 추가
